requirements.txt
requests==2.31.0
python-dotenv==1.0.0 
langchain==0.3.21
langchain-core==0.3.47
langchain-openai==0.3.9
supabase==2.14.0